requirements.txt
streamlit>=1.32.0
psycopg2-binary>=2.9.9
bcrypt>=4.1.2
pandas>=2.0.0
altair>=5.0.0
reportlab>=4.0.0
python-dotenv>=1.0.0
